// geremi-ihm/src/app/shared/enums/etape.enums.ts
export enum Etape {
  // ETUDE 
  CREATION = 'CREATION',
  ZONAGE = 'ZONAGE',
  POPULATION = 'POPULATION',
  CONTRAINTES = 'CONTRAINTES',
  CHANTIERS = 'CHANTIERS',
  INSTALLATIONS = 'INSTALLATIONS',
  MATERIAUX = 'MATERIAUX',
}

export enum EtapeScenario {
  // SCENARIO
  CREATION_SCENARIO = 'CREATION_SCENARIO',
  CONTRAINTES_SCENARIO = 'CONTRAINTES_SCENARIO',
  CHANTIERS_SCENARIO = 'CHANTIERS_SCENARIO',
  INSTALLATIONS_SCENARIO = 'INSTALLATIONS_SCENARIO',
  MATERIAUX_SCENARIO = 'MATERIAUX_SCENARIO',
  HYPOTHESE_VENTILATION_SCENARIO = 'HYPOTHESE_VENTILATION_SCENARIO',
  HYPOTHESE_PROJECTION_SCENARIO = 'HYPOTHESE_PROJECTION_SCENARIO'
}